+++ TripoliApp/src/main/resources/org/cirdles/tripoli/gui/settings/color/fxcomponents/ColorPickerSplotch.fxml
